backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.3,<5.6
django-cors-headers>=4.4,<4.8
django-filter>=24.2,<25.2
psycopg[binary]>=3.2,<3.3
redis>=5.2,<6
celery>=5.4,<5.6
psutil>=6.0,<7.1
waitress>=3.0,<3.1
